Dynamically generate shutter.qrc based on flavor
The app used to have a pre-generated shutter.qrc that pointed to
the platform.kirigami QML files independently of the chosen flavor.

As far as I can tell, this prevents the app from building correctly
for other flavors.

This commit replaces `shutter.qrc` with `shutter.qrc.in` and adds a
step in `harbour-shutter.pro` to replace the `@FLAVOR@` placeholder.

Signed-off-by: Sumwun Ducainest <sumwun.ducainest@outlook.com>