Add Ubuntu Touch libraries build
Shutter relies on 2 libraires that are unavailable in the Focal (20.04)
repos that Ubuntu Touch is currently still using:
  - libcamera
  - libopencv
For Ubuntu Touch, we therefore need to build AND ship those libraries
with the app.
This commit adds the clickable configuration to do that, as well as a
script that will be used to build libcamera.

Signed-off-by: Sumwun Ducainest <sumwun.ducainest@outlook.com>