Ubuntu Touch: Add QtMultimedia's qvideoframe_p.h
This commit adds the QtMultimedia qvideoframe_p.h private header to
the Ubuntu Touch-specific files.
This is necessary, because Qt no longer distributes private headers,
so the package containing the header is not available on Ubuntu.
In time, the use of this header should be replaced by the Qt-recommended
way of doing things.

Signed-off-by: Sumwun Ducainest <sumwun.ducainest@outlook.com>