Make face detection optional: Add a button
This commit adds a face detection toggle in the general panel of the settings
for the user to be able to turn face detection on and off at runtime.

Note that because the settings so far were on the QML side only, the setting for
this button currently lives in a different object.
Ultimately, we might want to merge both setting scopes into one singl shared one.

Signed-off-by: Sumwun Ducainest <sumwun.ducainest@outlook.com>